+++ tiger-test-lib/src/main/java/io/cucumber/core/plugin/ScenarioContextDelegate.java
@@ -16,13 +16,19 @@
 
 package io.cucumber.core.plugin;
 
+import lombok.Getter;
 import lombok.experimental.Delegate;
+
+import java.net.URI;
 
 public class ScenarioContextDelegate {
 
-  public ScenarioContextDelegate(ScenarioContext context) {
+  @Getter private final URI featureURI;
+
+  public ScenarioContextDelegate(URI featureURI, ScenarioContextParallel context) {
+    this.featureURI = featureURI;
     this.context = context;
   }
 
-  @Delegate private final ScenarioContext context;
+  @Delegate private final ScenarioContextParallel context;
 }